pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "aml-monitoring"
version = "0.1.0"
description = "AML Transaction Monitoring MVP for fintech security"
authors = ["Fintech Security Team"]
readme = "README.md"
packages = [{ include = "aml_monitoring", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
sqlalchemy = "^2.0.36"
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
typer = { extras = ["all"], version = "^0.12.0" }
pyyaml = "^6.0.2"
pandas = "^2.2.0"
aiosqlite = "^0.20.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.18.4"
streamlit = "^1.40.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
mypy = "^1.13.0"
ruff = "^0.8.0"
black = "^24.10.0"
pre-commit = "^4.0.0"
httpx = "^0.27.0"
alembic = "^1.14.0"

[tool.poetry.scripts]
aml = "aml_monitoring.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"
[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "def __repr__", "raise NotImplementedError"]
